+++ src/config/adwords/reports/AD_PERFORMANCE_REPORT/clicks.php
@@ -3,7 +3,7 @@ return [
     "id" => "clicks",
     "property" => "Clicks",
     "is_filter" => TRUE,
-    "type" => "long",
+    "type" => "integer",
     "is_metric" => TRUE,
     "is_dimension" => FALSE,
     "is_percentage" => FALSE
