+++ src/config/adwords/reports/AD_PERFORMANCE_REPORT/conversiontrackerid.php
@@ -3,7 +3,7 @@ return [
     "id" => "conversiontrackerid",
     "property" => "ConversionTrackerId",
     "is_filter" => TRUE,
-    "type" => "long",
+    "type" => "integer",
     "is_metric" => FALSE,
     "is_dimension" => TRUE,
     "is_percentage" => FALSE
