+++ src/config/adwords/reports/BUDGET_PERFORMANCE_REPORT/valueperallconversion.php
@@ -3,7 +3,7 @@ return [
     "id" => "valueperallconversion",
     "property" => "ValuePerAllConversion",
     "is_filter" => TRUE,
-    "type" => "double",
+    "type" => "decimal",
     "is_metric" => TRUE,
     "is_dimension" => FALSE,
     "is_percentage" => FALSE
