+++ src/config/adwords/reports/CAMPAIGN_PERFORMANCE_REPORT/averagetimeonsite.php
@@ -3,7 +3,7 @@ return [
     "id" => "averagetimeonsite",
     "property" => "AverageTimeOnSite",
     "is_filter" => TRUE,
-    "type" => "double",
+    "type" => "decimal",
     "is_metric" => TRUE,
     "is_dimension" => FALSE,
     "is_percentage" => FALSE
