+++ src/config/adwords/reports/KEYWORDS_PERFORMANCE_REPORT/baseadgroupid.php
@@ -3,7 +3,7 @@ return [
     "id" => "baseadgroupid",
     "property" => "BaseAdGroupId",
     "is_filter" => TRUE,
-    "type" => "long",
+    "type" => "integer",
     "is_metric" => FALSE,
     "is_dimension" => TRUE,
     "is_percentage" => FALSE
